requirements.txt
streamlit == 1.29.0
pandas == 2.0.1
scikit-learn == 1.2.2
nltk == 3.8.1
bs4 == 0.0.1
requests == 2.30.0
datetime == 5.2
regex == 2023.6.3
matplotlib == 3.7.1
seaborn == 0.12.2
lxml==4.6.3
youtube-search-python == 1.6.6
sqlalchemy == 2.0.24
dotenv == 1.0.0
mysql-connector-python == 8.2.0